requirements.txt
Flask==1.1.2
Keras==2.4.3
numpy==1.19.5
Pillow==8.3.1
tensorflow==2.4.0